# docs/openapi.json
{
  "components": {
    "schemas": {
      "BookmarkCreate": {
        "description": "Bookmark creation input.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "folder": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "source": {
            "$ref": "#/components/schemas/BookmarkSource",
            "default": "manual"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "title"
        ],
        "title": "BookmarkCreate",
        "type": "object"
      },
      "BookmarkPresentationPage": {
        "description": "Paginated bookmark presentation output.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookmarkPresentationRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "BookmarkPresentationPage",
        "type": "object"
      },
...
      "BookmarkRead": {
        "description": "Bookmark output.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "folder": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/BookmarkSource",
            "default": "manual"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "title",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "BookmarkRead",
        "type": "object"
      },
      "BookmarkSort": {
        "description": "Bookmark list sort order by original save time.",
        "enum": [
          "newest",
          "oldest"
        ],
        "title": "BookmarkSort",
        "type": "string"
      },
...
      "BookmarkUpdate": {
        "description": "Partial bookmark update input.",
        "properties": {
          "folder": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "BookmarkUpdate",
        "type": "object"
      },
      "ErrorDetail": {
        "description": "API error response.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Link Hoarder API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/bookmarks": {
      "get": {
        "operationId": "list_bookmarks_api_v1_bookmarks_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BookmarkSort"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkPresentationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Bookmarks",
        "tags": [
          "bookmarks"
        ]
      },
      "post": {
        "operationId": "create_bookmark_api_v1_bookmarks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Bookmark",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/v1/bookmarks/by-url": {
      "get": {
        "operationId": "get_bookmark_by_url_api_v1_bookmarks_by_url_get",
        "parameters": [
          {
            "in": "query",
            "name": "url",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Bookmark By Url",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/v1/bookmarks/{bookmark_id}": {
      "delete": {
        "operationId": "delete_bookmark_api_v1_bookmarks__bookmark_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "title": "Bookmark Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Bookmark",
        "tags": [
          "bookmarks"
        ]
      },
      "get": {
        "operationId": "get_bookmark_api_v1_bookmarks__bookmark_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "title": "Bookmark Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Bookmark",
        "tags": [
          "bookmarks"
        ]
      },
      "patch": {
        "operationId": "update_bookmark_api_v1_bookmarks__bookmark_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "title": "Bookmark Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookmarkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Bookmark",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/v1/bookmarks/{bookmark_id}/favicon": {
      "get": {
        "operationId": "get_bookmark_favicon_api_v1_bookmarks__bookmark_id__favicon_get",
        "parameters": [
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "title": "Bookmark Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Bookmark Favicon",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/v1/bookmarks/{bookmark_id}/thumbnail": {
      "get": {
        "operationId": "get_bookmark_thumbnail_api_v1_bookmarks__bookmark_id__thumbnail_get",
        "parameters": [
          {
            "in": "path",
            "name": "bookmark_id",
            "required": true,
            "schema": {
              "title": "Bookmark Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Bookmark Thumbnail",
        "tags": [
          "bookmarks"
        ]
      }
    },
    "/api/v1/imports/bookmarks-file": {
      "post": {
        "operationId": "import_bookmarks_file_api_v1_imports_bookmarks_file_post",
        "requestBody": {
          "content": {
            "text/html": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "maxLength": 16777216,
                "minLength": 1,
                "title": "Content",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Import Bookmarks File",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/bookmarks-json": {
      "post": {
        "operationId": "import_bookmarks_json_api_v1_imports_bookmarks_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/BookmarkCreate"
                },
                "maxItems": 50000,
                "minItems": 1,
                "title": "Bookmarks",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Import Bookmarks Json",
        "tags": [
          "imports"
        ]
      }
    },
...
  }
}